using double prec

+++ include/DecisionNode.hpp
@@ -7,7 +7,8 @@
 namespace Kaadugal
 {
 	// Use Variable Precision Floating point representation
-	typedef float VPFloat; // Change to float or long double as needed
+	//typedef float VPFloat; // Change to float or long double as needed
+	typedef double VPFloat; // Change to float or long double as needed
 	enum NodeType
 	{
 		SplitNode,
